Add files via upload
This master thesis investigates the influence of macro- and micronutrients on the human microbiota based on dietary data from a Food Frequency Questionnaire (FFQ) and microbial 16S rRNA sequencing data from stool samples of a healthy control cohort. The FFQ data were converted into food groups as well as macro- and micronutrients to compare both classifications in terms of their informative value for microbiological associations.
This shows the analysis of alpha diversity, beta diversity and MaAsLin of macro- and micronutrients.